Add Matomo for analytics
This commit adds back Matomo for analytics.
The instance is hosted by the Open Knowledge Foundation so
we can consider it privacy-consious.

I also made sure to add the line
```
_paq.push(['disableCookies']);
```
which disables setting a cookie on the users browser. This does
not allow us to track users across sesssions, what we mostly care
about though is the overall impressions - i.e. which pages do
people look at and when. We do not care about individual behaviour
at all.