Add Matomo for analytics (#213)
* Add Matomo for analytics

This commit adds back Matomo for analytics.
The instance is hosted by the Open Knowledge Foundation so
we can consider it privacy-consious.

I also made sure to add the line
```
_paq.push(['disableCookies']);
```
which disables setting a cookie on the users browser. This does
not allow us to track users across sesssions, what we mostly care
about though is the overall impressions - i.e. which pages do
people look at and when. We do not care about individual behaviour
at all.

* Move analytics to end of body to avoid blocking page load

* Bring back Matomo section in Datenschutz

This is the very same text that is used on okfn.de
which uses the same matomo instance.

* Remove Mapbox section from Datenschutz

In PR #212 we removed the tiles that we loaded
from mapbox and replaced them with the regular
openstreetmap.org tiles. Hence we can remove the
mapbox part of the Datenschutz page

* Disable option to hide footer

no page ever decided to hide the footer so
we agreed to remove this check to simplify 
the code