Fix image attribution display on large screens
This was noticed because the image that was uploaded for the last blogpost
has rather small intrinsic dimensions (width of 640px) and would not grow
to the full width of the page. The attribution would still be right aligned
though which looked odd.

By converting the `figure` to use `display: flex` we can ensure that the image
is displayed in full width and properly aligns with the attribution again.

Fixes #291