Hide leaflet tiles when taking percy screenshots
With almost every change that we push, percy shows us changes.
This is because part of our website is a map which renders
OpenStreetMap tiles. Whenever those tiles change, percy notices
that as a change (even though this is of course independent from
our actual website) and marks the screenshot as a change.

With this change, we hide all leaflet tiles so that they no longer
get displayed and hence cannot cause changes in our screenshots
anymore.