Replace deprecated percy-snapshot-action (#323)
* Replace deprecated percy-snapshot-action

As per the [readme](https://github.com/percy/snapshot-action), this
action was deprecated in favor of just running `npx percy` as in this
commit. A failing build on our side triggered this requirement. See
[build logs](https://github.com/okfde/codefor.de/actions/runs/1677073844).

* Update to new percy format

This was done using `percy migrate`. Further tweaks
were made to the files to be snapshotted.
It also required a move of the percy specific CSS as per the
[documentation](https://docs.percy.io/docs/percy-specific-css#percy-css-media-query)