Fix/percy static config (#350)
* Ensure percy only takes screenshots of approved pages

It seems like there was a change in how percy handles (relative)
paths which led to it taking screenshots of every single page
and draining our budget. This commit adapts the paths to the new
format so that only the 7 required screenshots are taken.

* Pin percy version in Github actions

This is so that we don't accidentally use a newer
version in the actions that run in Github which might
not be compatible with our config anymore.