Move ziele.md to themen.md and rewrite ziele.md
The idea was to focus on our goals "showing what's possible" and
demanding that "doing what's neccessary" is done by the government
(credits to Cyd Harrell for this distinction). Showing what's possible
is explained with the help of numerous examples to underline the amount
of work that has already been spent here. The article explains why many
projects died and concludes with demands on administration and politics
(work-in-progress).