Use design system colors in icons (#466)
The blue icon variant was previously achieved by using
a `filter` on the `img` which would not lead to quite the color that we wanted. Since these icons are pretty small SVGs, we can instead inline them and use the css `currentColor` field in their styles to pass in the proper color variable from the css.